// ems-core/src/shared/models/MatchState.ts
export enum MatchState {
  PRESTART_READY = "READY TO PRESTART",
  PRESTART_IN_PROGRESS = "PRESTARTING...",
  PRESTART_COMPLETE = "PRESTART COMPLETE",
  AUDIENCE_DISPLAY_SET = "AUDIENCE DISPLAY SET",
  MATCH_READY = "MATCH READY",
  MATCH_NOT_READY = "MATCH NOT READY",
  MATCH_IN_PROGRESS = "MATCH IN PROGRESS",
  MATCH_COMPLETE = "MATCH ENDED",
  MATCH_ABORTED = "MATCH ABORTED"
}